requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
fastapi==0.112.1
fastapi-utils==0.7.0
h11==0.14.0
idna==3.7
mypy-extensions==1.0.0
psutil==5.9.8
pydantic==2.8.2
pydantic_core==2.20.1
pymongo==4.8.0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
starlette==0.38.2
typing_extensions==4.12.2
typing-inspect==0.9.0
urllib3==2.2.2
uvicorn==0.30.6
